OpenPaaS-Suite/esn-frontend-mailto#35 Move directive opInboxCompose and its test to separate files
